requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.1.7
django-import-export==4.3.7
mysqlclient==2.2.7
pillow==11.1.0
python-dotenv==1.0.1
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.1
